updates to the backend

+++ backend/controllers/init_event_log.py
@@ -34,14 +34,25 @@ def _refresh_tree(workspaceData: WorkspaceData, cases, meta):
 
 def _import_event_log(log_path: str, workspace: Workspace):
   try:
-    workspace.clear_directory()
-    event_log = import_log(log_path)
-    event_log.to_pickle(workspace.get_file("log.pkl"))
-    cases, activities = generate_features(event_log)
+    workspaceData: WorkspaceData = workspace.data
+    activity_features = workspaceData.activity_features
+    del workspaceData
+    workspace.clear_directory_except_log()
+    if not os.path.exists(workspace.get_file("log.pkl")):
+      print("importing log" + log_path)
+      event_log = import_log(log_path)
+      event_log.to_pickle(workspace.get_file("log.pkl"))
+    else:
+      event_log = pd.read_pickle(workspace.get_file("log.pkl"))
+
+    
+    cases, activities, activity_features = generate_features(event_log, include_acivities=True, feature_map=activity_features)
     meta = describe_meta(cases)
 
-    workspaceData = workspace.data
+    workspaceData: WorkspaceData = workspace.data
+    workspaceData.activity_features = activity_features
     workspaceData.init_columns(meta)
+
     cache.clear()
     cache.workspace = workspaceData
     _refresh_tree(workspaceData, cases, meta)
@@ -52,17 +63,18 @@ def _import_event_log(log_path: str, workspace: Workspace):
     traceback.print_exc()
 
 async def import_event_log(job: Job):
-    log_name = job.data.get("log_name")
+    log_file = job.data.get("log_file")
     workspace: Workspace = job.data.get("workspace")
+    log_file = workspace.log_file if log_file is None else log_file
 
     logger = logging.getLogger('uvicorn')
     job.start_job()
     basepath = os.path.join(os.getcwd(), "lab", "data")
     
     logger.info(os.getcwd())
-    if log_name is None or log_name == "":
-      log_name = "bpi_challenge_2019_sample.xes"
-    log_path = os.path.join(basepath, log_name)
+    if log_file is None or log_file == "":
+      log_file = "bpi_challenge_2019_sample.xes"
+    log_path = os.path.join(basepath, log_file)
 
     loop = asyncio.get_running_loop()
